refactor code to use new eslint config
