+++ server/utils/constants.ts
@@ -1,4 +1,3 @@
-// @ts-ignore
 const ACCESS_DATA_COOKIE_NAME = 'accessData'
 
 module.exports = {
